Move example code for now
